reorganise files into "utils" and "settings" folders
