+++ src/settings/ui/snippets_editor/extensions.ts
